add a KPCA benchmark. Rename krill_simple.cc to krr.cc. Rename SEKLO to RBFKernelMatrix.
